# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Hub API",
        "description": "API para gerenciamento de eventos com venda de ingressos, processamento de pagamentos e notificações.",
        "contact": {
            "email": "contato@eventhub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Listar todos os eventos",
                "description": "Retorna uma lista de todos os eventos disponíveis",
                "operationId": "678607e594e433c58bfa0e62950b92ba",
                "responses": {
                    "200": {
                        "description": "Lista de eventos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "organizer": {
                                                        "$ref": "#/components/schemas/User"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Show de Rock"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Um grande show de rock"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2025-12-31 20:00:00"
                                                    },
                                                    "ticket_price": {
                                                        "type": "number",
                                                        "example": 50
                                                    },
                                                    "capacity": {
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "remaining_tickets": {
                                                        "type": "integer",
                                                        "example": 85
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Criar evento",
                "description": "Cria um novo evento. Requer autenticação de Organizador",
                "operationId": "c94193f1bc70d3546f62a20f9eb47d86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "date",
                                    "ticket_price",
                                    "capacity"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Show de Rock"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Um grande show de rock com bandas locais"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "ticket_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "organizer": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Show de Rock"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Um grande show de rock com bandas locais"
                                        },
                                        "date": {
                                            "type": "string",
                                            "example": "2025-12-31"
                                        },
                                        "ticket_price": {
                                            "type": "number",
                                            "example": 50
                                        },
                                        "capacity": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "remaining_tickets": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Não autorizado - apenas Organizadores podem criar eventos"
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Buscar evento por ID",
                "description": "Retorna os detalhes de um evento específico",
                "operationId": "2d12e555e58ea8f2fa8535f3303c6c87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do evento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do evento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "organizer": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Show de Rock"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Um grande show de rock"
                                        },
                                        "date": {
                                            "type": "string",
                                            "example": "2025-12-31 20:00:00"
                                        },
                                        "ticket_price": {
                                            "type": "number",
                                            "example": 50
                                        },
                                        "capacity": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "remaining_tickets": {
                                            "type": "integer",
                                            "example": 85
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Evento não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "type",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "ulid"
                    },
                    "name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "type": {
                        "type": "string",
                        "example": "ORGANIZER"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2025-10-20 12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2025-10-20 12:00:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via token Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}